backend/app/services/structured_api_service.py
"""
结构化 API 服务 - Mock 实现
用于获取精确的宏观经济数据和金融市场数据
"""
import logging
from typing import Dict, Any, Optional, List
from datetime import datetime, timedelta
import random

logger = logging.getLogger(__name__)


class StructuredAPIService:
    """
    结构化 API 服务
    当前阶段：所有 API 均为 Mock 实现，返回模拟数据
    未来阶段：逐步接入真实 API
    """
    
    def __init__(self):
        self.mock_mode = True
        logger.info("[StructuredAPI] 初始化（Mock 模式）")
    
    async def fetch_data(
        self, 
        api_name: str, 
        node_label: str,
        node_type: str,
        **kwargs
    ) -> Optional[Dict[str, Any]]:
        """
        统一的数据获取接口
        
        Args:
            api_name: API 名称 (fred, tushare, sec_edgar, ccxt, polygon)
            node_label: 节点标签
            node_type: 节点类型
            **kwargs: 额外参数
            
        Returns:
            {
                "value": "数值或状态",
                "unit": "单位",
                "timestamp": "数据时间戳",
                "source": "数据源",
                "metadata": {...}
            }
        """
        logger.info(f"[StructuredAPI] 调用 {api_name} 获取 {node_label} 数据")
        
        if api_name == "fred":
            return await self._fetch_fred(node_label, node_type)
        elif api_name == "tushare":
            return await self._fetch_tushare(node_label, node_type)
        elif api_name == "sec_edgar":
            return await self._fetch_sec_edgar(node_label, node_type)
        elif api_name == "ccxt":
            return await self._fetch_ccxt(node_label, node_type)
        elif api_name == "polygon":
            return await self._fetch_polygon(node_label, node_type)
        else:
            logger.warning(f"[StructuredAPI] 未知 API: {api_name}")
            return None
    
    async def _fetch_fred(self, node_label: str, node_type: str) -> Dict[str, Any]:
        """
        FRED API Mock
        
        TODO: 接入真实 FRED API
        - 申请 API Key: https://fred.stlouisfed.org/docs/api/api_key.html
        - 使用 requests 或 aiohttp 调用
        - 示例: https://api.stlouisfed.org/fred/series/observations?series_id=DFF&api_key=YOUR_KEY
        """
        logger.info(f"[FRED Mock] 获取 {node_label}")
        
        # Mock 数据映射
        mock_data = {
            "美联储利率": {
                "value": "5.25%-5.50%",
                "unit": "percentage",
                "series_id": "DFF",
                "description": "Federal Funds Effective Rate"
            },
            "美国通胀": {
                "value": "3.2%",
                "unit": "percentage",
                "series_id": "CPIAUCSL",
                "description": "Consumer Price Index for All Urban Consumers"
            },
            "美国失业率": {
                "value": "3.7%",
                "unit": "percentage",
                "series_id": "UNRATE",
                "description": "Unemployment Rate"
            },
            "美国GDP": {
                "value": "27.36万亿美元",
                "unit": "trillion_usd",
                "series_id": "GDP",
                "description": "Gross Domestic Product"
            }
        }
        
        # 模糊匹配
        for key, data in mock_data.items():
            if key in node_label or node_label in key:
                return {
                    "value": data["value"],
                    "unit": data["unit"],
                    "timestamp": (datetime.utcnow() - timedelta(days=1)).isoformat(),
                    "source": "FRED (Mock)",
                    "metadata": {
                        "series_id": data["series_id"],
                        "description": data["description"],
                        "api": "fred"
                    }
                }
        
        # 默认返回
        return {
            "value": f"{random.uniform(2.0, 5.0):.2f}%",
            "unit": "percentage",
            "timestamp": datetime.utcnow().isoformat(),
            "source": "FRED (Mock)",
            "metadata": {"api": "fred"}
        }
    
    async def _fetch_tushare(self, node_label: str, node_type: str) -> Dict[str, Any]:
        """
        Tushare API Mock
        
        TODO: 接入真实 Tushare Pro API
        - 注册获取 Token: https://tushare.pro/register
        - 安装: pip install tushare
        - 示例代码:
            import tushare as ts
            ts.set_token('YOUR_TOKEN')
            pro = ts.pro_api()
            df = pro.index_daily(ts_code='000001.SH')
        """
        logger.info(f"[Tushare Mock] 获取 {node_label}")
        
        mock_data = {
            "上证指数": {
                "value": "3245.18",
                "unit": "points",
                "ts_code": "000001.SH",
                "change_pct": "+1.23%"
            },
            "人民币汇率": {
                "value": "7.24",
                "unit": "CNY/USD",
                "ts_code": "USDCNY",
                "change_pct": "-0.15%"
            },
            "中国CPI": {
                "value": "0.2%",
                "unit": "percentage",
                "ts_code": "CPI",
                "change_pct": "+0.1%"
            },
            "社融规模": {
                "value": "3.2万亿元",
                "unit": "trillion_cny",
                "ts_code": "SOCIAL_FINANCING",
                "change_pct": "+5.6%"
            }
        }
        
        for key, data in mock_data.items():
            if key in node_label or node_label in key:
                return {
                    "value": data["value"],
                    "unit": data["unit"],
                    "timestamp": (datetime.utcnow() - timedelta(hours=16)).isoformat(),
                    "source": "Tushare (Mock)",
                    "metadata": {
                        "ts_code": data["ts_code"],
                        "change_pct": data["change_pct"],
                        "api": "tushare"
                    }
                }
        
        return {
            "value": f"{random.uniform(3000, 3500):.2f}",
            "unit": "points",
            "timestamp": datetime.utcnow().isoformat(),
            "source": "Tushare (Mock)",
            "metadata": {"api": "tushare"}
        }
    
    async def _fetch_sec_edgar(self, node_label: str, node_type: str) -> Dict[str, Any]:
        """
        SEC EDGAR API Mock
        
        TODO: 接入真实 SEC EDGAR API
        - 官方文档: https://www.sec.gov/edgar/sec-api-documentation
        - 需要设置 User-Agent 头
        - 示例: https://data.sec.gov/submissions/CIK0000320193.json (Apple)
        """
        logger.info(f"[SEC EDGAR Mock] 获取 {node_label}")
        
        return {
            "value": "最新 10-K 已披露",
            "unit": "filing_status",
            "timestamp": (datetime.utcnow() - timedelta(days=7)).isoformat(),
            "source": "SEC EDGAR (Mock)",
            "metadata": {
                "filing_type": "10-K",
                "cik": "0000320193",
                "api": "sec_edgar"
            }
        }
    
    async def _fetch_ccxt(self, node_label: str, node_type: str) -> Dict[str, Any]:
        """
        CCXT API Mock
        
        TODO: 接入真实 CCXT
        - 安装: pip install ccxt
        - 示例代码:
            import ccxt
            exchange = ccxt.binance()
            ticker = exchange.fetch_ticker('BTC/USDT')
        """
        logger.info(f"[CCXT Mock] 获取 {node_label}")
        
        mock_data = {
            "比特币": {
                "value": "43250.00",
                "unit": "USDT",
                "symbol": "BTC/USDT",
                "change_24h": "+2.34%"
            },
            "以太坊": {
                "value": "2280.50",
                "unit": "USDT",
                "symbol": "ETH/USDT",
                "change_24h": "+1.87%"
            }
        }
        
        for key, data in mock_data.items():
            if key in node_label or node_label in key:
                return {
                    "value": data["value"],
                    "unit": data["unit"],
                    "timestamp": datetime.utcnow().isoformat(),
                    "source": "CCXT (Mock)",
                    "metadata": {
                        "symbol": data["symbol"],
                        "change_24h": data["change_24h"],
                        "api": "ccxt"
                    }
                }
        
        return {
            "value": f"{random.uniform(40000, 50000):.2f}",
            "unit": "USDT",
            "timestamp": datetime.utcnow().isoformat(),
            "source": "CCXT (Mock)",
            "metadata": {"api": "ccxt"}
        }
    
    async def _fetch_polygon(self, node_label: str, node_type: str) -> Dict[str, Any]:
        """
        Polygon.io API Mock
        
        TODO: 接入真实 Polygon.io API
        - 申请 API Key: https://polygon.io/
        - 示例: https://api.polygon.io/v2/aggs/ticker/AAPL/prev?apiKey=YOUR_KEY
        """
        logger.info(f"[Polygon Mock] 获取 {node_label}")
        
        return {
            "value": f"{random.uniform(150, 200):.2f}",
            "unit": "USD",
            "timestamp": datetime.utcnow().isoformat(),
            "source": "Polygon.io (Mock)",
            "metadata": {
                "ticker": "AAPL",
                "api": "polygon"
            }
        }

