backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.54.0
httpx==0.27.0
python-multipart==0.0.6
aiohttp==3.9.1
duckduckgo-search==4.1.0
yfinance==0.2.40

